Sentiment-Analysis/requirements.txt
numpy
pandas
pandas-datareader
sklearn
tensorflow
nltk
mysql-connector-python
praw
matplotlib
squarify